Fixes to RK31 to allow FRAMEBUFFER_CONSOLE (rk3188_lcdc and rk_fb)
This is basically the same fix that rk3066 needed to boot with this
kernel, now applied to rk3188_lcdc.c too.
Problem arises because RK moved the screen turning on to lcdc removing
it from rk_fb, and then the framebuffer_console fails (producing a stick
reboot) for not finding the screen.

Recommended to set CONFIG_FRAMEBUFFER_CONSOLE and unset CONFIG_LOGO or
else only half the screen is available for the boot log