dm9601: support dm9620 variant.
*** Copied from kernel 3.10:
 dm9620 is a newer variant of dm9601 with more features (usb 2.0, checksum
offload, ..), but it can also be put in a dm9601 compatible mode, allowing
us to reuse the existing driver.

This does mean that the extended features like checksum offload cannot be
used, but that's hardly critical on a 100mbps interface.
***

This is kind of a backport from linux kernel 3.10. Instead of identify
device type by reading device id, this change specify device type explicitly.

A few dm9601 and dm9620 devices were tested working with this change. As
before the change, IP addr can not be obtained through dhclient except a few
rare times. Static IP addr works flawlessly.

With this change it's better unselect the existing dm9620 driver in
"make menuconfig".